+++ .gitignore
@@ -1,5 +1,9 @@
-/env    
+venv/    
 /vscode
 .vscode
 __pycache__/
+htmlcov/
 .pytest_cache/
+__init__.pyc
+*.pyc
+.cache/
