+++ tests/test_app.py
@@ -15,6 +15,7 @@ class TestUserEndpoints(unittest.TestCase):
         response = self.myapp.post(self.user_endpoint)
         self.assertEqual(response.status_code, 200)
         
+        
     def test_list_all_questions(self):
         response = self.myapp.get(self.user_endpoint)
         self.assertEquals(response.status_code, 200)
